Move new code to _extra files
