+++ internal/ethapi/api_test.go
@@ -62,9 +62,7 @@ import (
 	"github.com/ethereum/go-ethereum/ethdb"
 	"github.com/ethereum/go-ethereum/event"
 	"github.com/holiman/uint256"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.uber.org/mock/gomock"
 	"golang.org/x/exp/slices"
 )
 
@@ -2066,103 +2064,4 @@ func testRPCResponseWithFile(t *testing.T, testid int, result interface{}, rpc s
 		t.Fatalf("error reading expected test file: %s output: %v", outputFile, err)
 	}
 	require.JSONEqf(t, string(want), string(data), "test %d: json not match, want: %s, have: %s", testid, string(want), string(data))
-}
-
-func TestBlockChainAPI_isAllowedToBeQueried(t *testing.T) {
-	t.Parallel()
-
-	const recentBlocksWindow uint64 = 1024
-
-	makeBlockWithNumber := func(number uint64) *types.Block {
-		header := &types.Header{
-			Number: big.NewInt(int64(number)),
-		}
-		return types.NewBlock(header, nil, nil, nil, nil)
-	}
-
-	testCases := map[string]struct {
-		blockNumOrHash rpc.BlockNumberOrHash
-		makeBackend    func(ctrl *gomock.Controller) *MockBackend
-		wantErrMessage string
-	}{
-		"historical_proofs_accepted": {
-			blockNumOrHash: rpc.BlockNumberOrHashWithNumber(rpc.BlockNumber(0)),
-			makeBackend: func(ctrl *gomock.Controller) *MockBackend {
-				backend := NewMockBackend(ctrl)
-				backend.EXPECT().HistoricalConfig().Return(true, recentBlocksWindow)
-				return backend
-			},
-		},
-		"block_number_recent_below_window": {
-			blockNumOrHash: rpc.BlockNumberOrHashWithNumber(rpc.BlockNumber(1000)),
-			makeBackend: func(ctrl *gomock.Controller) *MockBackend {
-				backend := NewMockBackend(ctrl)
-				backend.EXPECT().HistoricalConfig().Return(false, recentBlocksWindow)
-				backend.EXPECT().LastAcceptedBlock().Return(makeBlockWithNumber(1020))
-				return backend
-			},
-		},
-		"block_number_recent": {
-			blockNumOrHash: rpc.BlockNumberOrHashWithNumber(rpc.BlockNumber(2000)),
-			makeBackend: func(ctrl *gomock.Controller) *MockBackend {
-				backend := NewMockBackend(ctrl)
-				backend.EXPECT().HistoricalConfig().Return(false, recentBlocksWindow)
-				backend.EXPECT().LastAcceptedBlock().Return(makeBlockWithNumber(2200))
-				return backend
-			},
-		},
-		"block_number_recent_by_hash": {
-			blockNumOrHash: rpc.BlockNumberOrHashWithHash(common.Hash{99}, false),
-			makeBackend: func(ctrl *gomock.Controller) *MockBackend {
-				backend := NewMockBackend(ctrl)
-				backend.EXPECT().HistoricalConfig().Return(false, recentBlocksWindow)
-				backend.EXPECT().LastAcceptedBlock().Return(makeBlockWithNumber(2200))
-				backend.EXPECT().
-					BlockByNumberOrHash(gomock.Any(), gomock.Any()).
-					Return(makeBlockWithNumber(2000), nil)
-				return backend
-			},
-		},
-		"block_number_recent_by_hash_error": {
-			blockNumOrHash: rpc.BlockNumberOrHashWithHash(common.Hash{99}, false),
-			makeBackend: func(ctrl *gomock.Controller) *MockBackend {
-				backend := NewMockBackend(ctrl)
-				backend.EXPECT().HistoricalConfig().Return(false, recentBlocksWindow)
-				backend.EXPECT().LastAcceptedBlock().Return(makeBlockWithNumber(2200))
-				backend.EXPECT().
-					BlockByNumberOrHash(gomock.Any(), gomock.Any()).
-					Return(nil, fmt.Errorf("test error"))
-				return backend
-			},
-			wantErrMessage: "getting block from hash: test error",
-		},
-		"block_number_historical": {
-			blockNumOrHash: rpc.BlockNumberOrHashWithNumber(rpc.BlockNumber(1000)),
-			makeBackend: func(ctrl *gomock.Controller) *MockBackend {
-				backend := NewMockBackend(ctrl)
-				backend.EXPECT().HistoricalConfig().Return(false, recentBlocksWindow)
-				backend.EXPECT().LastAcceptedBlock().Return(makeBlockWithNumber(2200))
-				return backend
-			},
-			wantErrMessage: "block number 1000 is before the oldest non-historical block number 1176 (window of 1024 blocks)",
-		},
-	}
-
-	for name, testCase := range testCases {
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			ctrl := gomock.NewController(t)
-
-			api := &BlockChainAPI{
-				b: testCase.makeBackend(ctrl),
-			}
-
-			err := api.isAllowedToBeQueried(testCase.blockNumOrHash)
-			if testCase.wantErrMessage == "" {
-				assert.NoError(t, err)
-			} else {
-				assert.EqualError(t, err, testCase.wantErrMessage)
-			}
-		})
-	}
 }
